requirements.txt
streamlit==1.45.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.1
pyarrow==13.0.0
altair==5.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
imageio==2.31.1
requests==2.31.0
streamlit-nested-layout==0.1.4
streamlit-pdf-viewer==0.0.23
opencv-python==4.8.1.78
tifffile==2023.7.10
streamlit-image-coordinates==0.1.6
